requirements.txt
requests
beautifulsoup4
pandas
numpy
scikit-learn
nltk
sentence-transformers
typer[all]
python-dotenv
pyyaml
jsonschema
spacy
google-generativeai
